+++ CHANGELOG.md
@@ -74,6 +74,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **movie_clean_token** UI widget had a bug when selecting/deleting the top most item would result in a prompt
   to ask the user if they'd like to reset to default over and over again.
 - Modifying tracker settings in the Tracker widget page will now update the Tracker widget in the Tracker settings tab.
+- Bug when utilizing Plugin mode utilizing the built in Basic profile, could result in incorrect image generation being done.
 
 ### Removed